requirements.txt
Flask
Flask-CORS
tensorflow
keras
numpy
opencv-python
Pillow
werkzeug
gunicorn
python-multipart
